feat: Add new filters for additional criteria
While the hard-coded user meta key is handy, there may be other criteria that are desired to hide user profile pages and user profiles from members directory.

This commit consolidates some logic, adds a couple of new filters, and updates the documentation with examples of how to use the new filters.